Use the first occurence of a parameter default-value [motiondsl]
That is, if a parameter appears multiple times in a document, every
corresponding instance will get the same default value, equal to the
first value that appears in the model, if any.

This is to allow the author of the document to write the default
value only once per parameter, no matter how many times the parameter
appears in the motion steps.