Make sure evalf() returns a native float [!]
... and not a Sympy.core.numbers.Float instance.

When evaluating a constant Expression. To avoid surprises in
client code, and because the docs talk about 'float'.

E.g. this fixes the test 'test_pi_expr_value' which would
occasionally (!!) fail, because Sympy would return
either Float(0.0) or ...numbers.Zero() - :facepalm:

A potentially dangerous change, as it is deep in a
core module used by other projects...